==> src/01-LoadPackages.R <==
#Load packages
install.packages("pacman")
pacman::p_load(pacman, tidyverse, RColorBrewer, grid, ggpubr, here, vegan, 
               data.table, jpeg, tiff, DESeq2, readr, ggrepel, EnhancedVolcano,
               rstatix, phyloseq, microbiome, pairwiseAdonis, ggh4x)
